move mvt related function to mvt.go
